refactor(mergeMap): move support types under mergeMap.ts
